pyproject.toml
[project]
name = "sih-problem-edna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
]
